refactor(daemon): extract memory synthesis bridge
